[core] remove leftover launch and completion analytics from UpdateChecker (#13918)
* remove leftover launch and completion analytics

* split tests
