+++ fastlane_core/lib/fastlane_core/update_checker/update_checker.rb
@@ -17,14 +17,6 @@ module FastlaneCore
 
       Thread.new do
         begin
-          send_launch_analytic_events_for(gem_name)
-        rescue
-          # we don't want to show a stack trace if something goes wrong
-        end
-      end
-
-      Thread.new do
-        begin
           server_results[gem_name] = fetch_latest(gem_name)
         rescue
           # we don't want to show a stack trace if something goes wrong
@@ -46,14 +38,6 @@ module FastlaneCore
     end
 
     def self.show_update_status(gem_name, current_version)
-      Thread.new do
-        begin
-          send_completion_events_for(gem_name)
-        rescue
-          # we don't want to show a stack trace if something goes wrong
-        end
-      end
-
       if update_available?(gem_name, current_version)
         show_update_message(gem_name, current_version)
       end
@@ -120,179 +104,6 @@ module FastlaneCore
 
     def self.generate_fetch_url(gem_name)
       "https://rubygems.org/api/v1/gems/#{gem_name}.json"
-    end
-
-    def self.send_launch_analytic_events_for(gem_name)
-      return if FastlaneCore::Env.truthy?("FASTLANE_OPT_OUT_USAGE")
-      ci = Helper.ci?.to_s
-      app_id_guesser = FastlaneCore::AppIdentifierGuesser.new(args: ARGV, gem_name: gem_name)
-      project_hash = app_id_guesser.p_hash
-      p_hash = project_hash if project_hash
-      platform = @platform if @platform # this has to be called after `p_hash`
-
-      send_launch_analytic_events(p_hash, gem_name, platform, ci)
-    end
-
-    def self.send_launch_analytic_events(p_hash, tool, platform, ci)
-      timestamp_seconds = Time.now.to_i
-
-      analytics = []
-      analytics << event_for_p_hash(p_hash, tool, platform, timestamp_seconds) if p_hash
-      analytics << event_for_launch(tool, ci, timestamp_seconds)
-
-      send_events(analytics)
-    end
-
-    def self.event_for_p_hash(p_hash, tool, platform, timestamp_seconds)
-      {
-        event_source: {
-          oauth_app_name: oauth_app_name,
-          product: 'fastlane'
-        },
-        actor: {
-          name: 'project',
-          detail: p_hash
-        },
-        action: {
-          name: 'update_checked'
-        },
-        primary_target: {
-          name: 'tool',
-          detail: tool || 'unknown'
-        },
-        secondary_target: {
-          name: 'platform',
-          detail: secondary_target_string(platform || 'unknown')
-        },
-        millis_since_epoch: timestamp_seconds * 1000,
-        version: 1
-      }
-    end
-
-    def self.event_for_launch(tool, ci, timestamp_seconds)
-      {
-        event_source: {
-          oauth_app_name: oauth_app_name,
-          product: 'fastlane'
-        },
-        actor: {
-          name: 'tool',
-          detail: tool || 'unknown'
-        },
-        action: {
-          name: 'launched'
-        },
-        primary_target: {
-          name: 'ci',
-          detail: ci
-        },
-        secondary_target: {
-          name: 'launch',
-          detail: secondary_target_string('')
-        },
-        millis_since_epoch: timestamp_seconds * 1000,
-        version: 1
-      }
-    end
-
-    def self.send_completion_events_for(gem_name)
-      return if FastlaneCore::Env.truthy?("FASTLANE_OPT_OUT_USAGE")
-
-      ci = Helper.ci?.to_s
-      install_method = if Helper.rubygems?
-                         'gem'
-                       elsif Helper.bundler?
-                         'bundler'
-                       elsif Helper.mac_app?
-                         'mac_app'
-                       elsif Helper.contained_fastlane?
-                         'standalone'
-                       elsif Helper.homebrew?
-                         'homebrew'
-                       else
-                         'unknown'
-                       end
-      duration = (Time.now - start_time).to_i
-      timestamp_seconds = Time.now.to_i
-
-      send_completion_events(gem_name, ci, install_method, duration, timestamp_seconds)
-    end
-
-    def self.send_events(analytics)
-      analytic_event_body = { analytics: analytics }.to_json
-
-      url = ENV["FASTLANE_METRICS_URL"] || "https://fastlane-metrics.fabric.io/public"
-      Excon.post(url,
-                 body: analytic_event_body,
-                 headers: { "Content-Type" => 'application/json' })
-    end
-
-    def self.event_for_completion(tool, ci, duration, timestamp_seconds)
-      {
-        event_source: {
-          oauth_app_name: oauth_app_name,
-          product: 'fastlane'
-        },
-        actor: {
-          name: 'tool',
-          detail: tool || 'unknown'
-        },
-        action: {
-          name: 'completed_with_duration'
-        },
-        primary_target: {
-          name: 'duration',
-          detail: duration.to_s
-        },
-        secondary_target: {
-          name: 'ci',
-          detail: secondary_target_string(ci)
-        },
-        millis_since_epoch: timestamp_seconds * 1000,
-        version: 1
-      }
-    end
-
-    def self.event_for_install_method(tool, ci, install_method, timestamp_seconds)
-      {
-        event_source: {
-          oauth_app_name: oauth_app_name,
-          product: 'fastlane'
-        },
-        actor: {
-          name: 'tool',
-          detail: tool || 'unknown'
-        },
-        action: {
-          name: 'completed_with_install_method'
-        },
-        primary_target: {
-          name: 'install_method',
-          detail: install_method
-        },
-        secondary_target: {
-          name: 'ci',
-          detail: secondary_target_string(ci)
-        },
-        millis_since_epoch: timestamp_seconds * 1000,
-        version: 1
-      }
-    end
-
-    def self.secondary_target_string(string)
-      return string
-    end
-
-    def self.oauth_app_name
-      return 'fastlane-refresher'
-    end
-
-    def self.send_completion_events(tool, ci, install_method, duration, timestamp_seconds)
-      analytics = []
-      analytics << event_for_completion(tool, ci, duration, timestamp_seconds)
-      analytics << event_for_install_method(tool, ci, install_method, timestamp_seconds)
-
-      send_events(analytics)
     end
   end
 end
